+++ R/sentiments.R
@@ -26,3 +26,44 @@
 #'  \item \url{http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010}
 #'  }
 "sentiments"
+
+
+#' Get a tidy data frame of a single sentiment lexicon
+#'
+#' Get specific sentiment lexicons in a tidy format, with one row per word,
+#' in a form that can be joined with a one-word-per-row dataset.
+#' Each of these comes from the included \code{\link{sentiments}} data frame,
+#' but this performs the filtering for a specific lexicon, and removes
+#' columns that are not used in that lexicon.
+#'
+#' @return A tbl_df with a \code{word} column, and either a \code{sentiment}
+#' column (if \code{lexicon} is "bing" or "nrc") or a numeric \code{score} column
+#' (if \code{lexicon} is "afinn").
+#'
+#' @examples
+#'
+#' library(dplyr)
+#' get_sentiments("afinn")
+#' get_sentiments("bing")
+#'
+#' @export
+get_sentiments <- function(lexicon = c("afinn", "bing", "nrc")) {
+  lex = match.arg(lexicon)
+
+  if (lex == "afinn") {
+    # turn uppercase: reverse compatibility issue
+    lex <- "AFINN"
+  }
+
+  ret <- sentiments %>%
+    dplyr::filter(lex == lexicon) %>%
+    dplyr::select(-lexicon)
+
+  if (lex == "AFINN") {
+    ret$sentiment <- NULL
+  } else {
+    ret$score <- NULL
+  }
+
+  ret
+}